refactor(storage): address review — core/files measureUsage, Result errors, decomposed operations
- Move filesystem measurement from packages/core/src/storage into a generic
  measureUsage method on the core/files IFileSystem abstraction; remove the
  core storage domain entirely
- Route task-lifecycle-utils through @emdash/core/files and @emdash/core/exec
  instead of raw node:fs / node:child_process
- Return Result with typed errors for expected worktree-cleanup failures
  instead of throwing
- Decompose storage-service into task-storage-rows helper and
  list-task-storage-usage / delete-storage-tasks operations
- Emit taskDeleted per successful deletion in batch deletes so a partial
  failure no longer resurrects deleted tasks in the renderer
