+++ packages/core/src/files/fs/file-system.ts
@@ -6,12 +6,14 @@ import { realpathOrResolve } from '../../watch';
 import { enumerate as enumerateFiles } from '../enumerate';
 import { classifyFileError, isFileNotFoundCode, type FileError } from '../errors';
 import { validateAbsolutePath } from '../paths';
+import { measureUsage } from './measure-usage';
 import type {
   FileEnumeration,
   FileEnumerationOptions,
   FileGlob,
   FileGlobOptions,
   FileStat,
+  FileUsage,
   IFileSystem,
   ReadBytesResult,
   ReadFileOptions,
@@ -106,6 +108,17 @@ export class FileSystem implements IFileSystem {
         ctime: stat.ctime,
         mode: stat.mode,
       });
+    } catch (error) {
+      return err(classifyFileError(error, validated.data));
+    }
+  }
+
+  async measureUsage(absPath: string): Promise<Result<FileUsage, FileError>> {
+    const validated = validateAbsolutePath(absPath);
+    if (!validated.success) return validated;
+
+    try {
+      return ok(await measureUsage(validated.data));
     } catch (error) {
       return err(classifyFileError(error, validated.data));
     }
