+++ packages/core/src/files/fs/index.ts
@@ -5,6 +5,8 @@ export type {
   FileGlob,
   FileGlobOptions,
   FileStat,
+  FileUsage,
+  FileUsageError,
   IFileSystem,
   ReadBytesResult,
   ReadFileOptions,
