+++ packages/core/src/files/types.ts
@@ -35,6 +35,8 @@ export type {
   FileGlob,
   FileGlobOptions,
   FileStat,
+  FileUsage,
+  FileUsageError,
   IFileSystem,
   ReadBytesResult,
   ReadFileOptions,
