--- packages/core/src/storage/index.ts
@@ -1,2 +1,0 @@
-export { measureTaskStorage } from './measurement';
-export type { PathStorageUsage, StorageScanError, StorageScanErrorType } from './types';
