--- packages/core/src/storage/types.ts
@@ -1,16 +1,0 @@
-export type StorageScanErrorType = 'not-found' | 'not-directory' | 'read-failed' | 'stat-failed';
-
-export type StorageScanError = {
-  type: StorageScanErrorType;
-  path: string;
-  message: string;
-};
-
-export type PathStorageUsage = {
-  path: string;
-  exists: boolean;
-  isDirectory: boolean;
-  apparentBytes: number;
-  reclaimableBytes: number;
-  errors: StorageScanError[];
-};
